requirements.txt
Flask==2.3.3
torch==2.0.1
numpy==1.24.3
python-dotenv==1.0.0
Werkzeug==2.3.7
Jinja2==3.1.2
pytest==7.4.0
hypothesis==6.82.0
scipy==1.11.2
